pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepface>=0.0.93",
    "django>=5.2",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "docxtpl>=0.19.1",
    "openpyxl>=3.1.5",
    "shapely>=2.1.0",
    "tf-keras>=2.19.0",
    "torch>=2.7.0",
]
